Allow for adding cart items with zero quantity. This will enable some business cases that are governed by some specific business rules.
I have several use-cases where the quantity is only added later via an update for specific usability/process, business and security rules. E.g. tickets for events are added to the cart but the people (quantity) is only done during the checkout process and based on specific business rules.

With the older version it was possible to update the cart item's quantity to 0 and the cart still returned the item when called with the content() method and pricing, etc. correctly which was awesome! :) Please consider to allow this again as the current lock down of quantity must be 1 is problematic.

Thanks!